feat: Quickly explore possibilities with staggered guesses
My first version artificially staggered the first guess, but other
guesses often used the same color multiple times in a row because of the
ordering produced by allStates. Staggering the pegs by 1 as allStates is
produced yields more staggered guesses, which tend to narrow down to
wins in fewer turns.